LMSC-65236 New Account changes for current month has been completed

+++ src/main/java/com/unisys/trans/cps/middleware/models/response/AgentResponseDTO.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class AgentResponseDTO {
 
-    private Long newAccount;
+    private int newAccount;
 
     List<TopAgentsResponseDTO>  topAgentsResponseDTOList;
